Event sourcing - multi-type Kafka Listener

+++ movie-review/src/main/java/me/mzalietin/imdbproject/moviereview/infrastructure/repo/MovieReviewDao.java
@@ -3,6 +3,8 @@ package me.mzalietin.imdbproject.moviereview.infrastructure.repo;
 import me.mzalietin.imdbproject.moviereview.domain.model.MovieReview;
 import me.mzalietin.imdbproject.moviereview.domain.model.MovieReviewKey;
 import me.mzalietin.imdbproject.moviereview.domain.service.spi.MovieReviewDataAccess;
+import me.mzalietin.imdbproject.moviereview.domain.service.spi.ResourceAlreadyExistsException;
+import me.mzalietin.imdbproject.moviereview.domain.service.spi.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -16,12 +18,31 @@ public class MovieReviewDao implements MovieReviewDataAccess {
     }
 
     @Override
-    public void save(final MovieReview r) {
-        repo.save(new MovieReviewEntity(r.username(), r.movieId(), r.rating(), r.comment()));
+    public void create(final MovieReview review) throws ResourceAlreadyExistsException {
+        var key = new MovieReviewKey(review.username(), review.movieId());
+        if (repo.existsById(key)){
+            throw new ResourceAlreadyExistsException("Review Key = " + key);
+        } else {
+            repo.save(new MovieReviewEntity(review.username(), review.movieId(), review.rating(), review.comment()));
+        }
     }
 
     @Override
-    public void delete(final MovieReviewKey key) {
-        repo.deleteById(key);
+    public void update(final MovieReview review) throws ResourceNotFoundException {
+        var key = new MovieReviewKey(review.username(), review.movieId());
+        if (repo.existsById(key)){
+            repo.save(new MovieReviewEntity(review.username(), review.movieId(), review.rating(), review.comment()));
+        } else {
+            throw new ResourceNotFoundException("Review Key = " + key);
+        }
+    }
+
+    @Override
+    public void delete(final MovieReviewKey key) throws ResourceNotFoundException {
+        if (repo.existsById(key)){
+            repo.deleteById(key);
+        } else {
+            throw new ResourceNotFoundException("Review Key = " + key);
+        }
     }
 }
